Feature: use gitignore templates from local repo directly
The gitignore templates used here is from dvcs/gitignore, which is
the same one used by gitignore.io.

Environment variable `FORGIT_GI_USE_LOCAL` is used as a switch to
enable this feature, or the original method is used by default.

`forgit::ignore::update` and `forgit::ignore::get` are rewritten
to download the repo, to get templates from the local repo.
local variable `preview` in `forgit:ignore` is also rewritten to
get the preview from there.


Former-commit-id: c1a8bd71dc771bc607acf866b7e209f0955e4dcd