Handle .gitconfig pager attribute.
The existing script handles the `core.pager` attribute, but does not
account for the `pager` attribute in git-config. Tools like
`diff-highlight` which ship with git recommend using the `pager`
attribute, and provide beautiful diffs.

I've made minor changes in the script to use these configuration
parameters if they are defined.

Ref: https://github.com/git/git/tree/master/contrib/diff-highlight