Implement git-checkout to switch branches in a repo
I added `gcb` which is used to list all the branches and checkout into
one of your choice. I believe this is a useful functionality especially
when a repo has too many branches to rememmber and lookup.

The preview shows the log graph from the selected branch.
I have removed keybinding for copy commit hash on gcb alias

Signed-off-by: b10n1k <jbonatakis@gmail.com>