Improve handling of arguments in forgit::add
currently, passing arguments to forgit::add bypasses the interactive selector and are passed directly to `git add`. This isn't particularly useful. So pass the arguments to the `git status` call instead, so fzf can be used to interactively select from the list.

This is particularly useful when calling `git forgit add` against a subdirectory or a shell glob (e.g. `git forgit add *(om[1,5])`)