Fix empty diff preview on renames
This had been fixed before in #189, but obviously the patch broke
support for whitespaces in file names. That was fixed in #204, which
in turn broke support for renames again.

Unfortunately we cannot support both cases together with the existing
'xargs' implementation, because the arguments are interpreted either
as one single file name (which breaks renames, because we have two files
passed in this case) or as separate file names (which breaks whitespace
support, because git would interpret each word of the file name as a
separate file).

This patch moves away from xargs and stores the file names in a bash
array instead. This makes it possible to pass them to "git diff" quoted
one by one using the "${array[@]}" notation.