feat: Use fzf-tmux if fzf options are set
The meat of this change is to inspect whether or not we are
in a `tmux` pane, and if so, we check the `FZF_TMUX` and
`FZF_TMUX_OPTS` variables. If those are set, we use `fzf-tmux`
instead of `fzf` and we forward the appropriate options.

- Add a `do_fzf` function for abstracting fzf invocation based on
shell state as outlined above.

- Add a new variable, FORGIT_SHOW_FZF_ON_EMPTY_LIST, for controlling when `--exit-0`
is forwarded to `fzf`.

- Replace all invocations of `fzf` with `do_fzf`.